- creates new GameStateHelpers hook - uses this hook to properly update the game state for rotations
